compile error

+++ src/OrbiterGen/Renderer.cpp
@@ -1298,7 +1298,7 @@ PlutoSize Renderer::RealRenderText(RendererImage * pRenderImage, DesignObjText *
 	try
 	{
 		//BOM (Byte Order Mask) header for UTF-8 strings
-		const char BOM_header[3] = { (BYTE)0xef, (BYTE)0xbb, (BYTE)0xbf };
+		const char BOM_header[3] = { 0xef, 0xbb, 0xbf };
 
 		//already a UTF-8 string?
 		bool bAlreadyHasBOMHeader = !memcmp(BOM_header, pDesignObjText->m_sText.data(), sizeof(BOM_header));
